Revert "making mean-path-length use single-source shortest path calls rather than shortest-path calls for each pair individually"
This reverts commit 999e055d73fa439c1baa2a26a106c2b3649704e6.

Jung's getDistanceMap(source) only gives you distances to actually reachable turtles, and the rest of the algorithm is relying on distance being null in case of non-reachability to ultimately return false. The mean-path-length-directed-two-half-connected-turtles was already catching that, and I am going top push a new mean-path-length-two-connected-and-one-unconnected-turtles test to make it even clearer.